% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan_ivreg.R
\name{print.stan_ivreg}
\alias{print.stan_ivreg}
\title{print.stan_ivreg a print method for stan_ivreg}
\usage{
\method{print}{stan_ivreg}(model_fit)
}
\arguments{
\item{model_fit}{a model fit by}
}
\description{
print.stan_ivreg a print method for stan_ivreg
}

