% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan_ivreg.R
\name{stan_iv}
\alias{stan_iv}
\title{Runs Bayesian IV}
\usage{
stan_iv(Y, X, Z, ...)
}
\arguments{
\item{Y}{a vector of outcomes}

\item{Z}{a matrix of instruments}

\item{...}{additional arguments to be passed to rstan::sampling}

\item{X_endog}{a matrix of endogenous regressors}

\item{offset}{a vector offset}
}
\description{
Runs Bayesian IV
}

